+++ docs/voluseg-docs-app/docs/getting_started/running.mdx
@@ -25,7 +25,7 @@ To run the demo you will need start by downloading the sample data to your local
 import voluseg
 
 # Download sample data
-voluseg._tools.download_sample_data("/path/to/input/")
+voluseg.tools.download_sample_data("/path/to/input/")
 ```
 
 ### Setup pipeline parameters
