+++ src/voluseg/__init__.py
@@ -1,14 +1,14 @@
 """initialization"""
 
-from voluseg._steps.step0 import process_parameters as step0_process_parameters
-from voluseg._steps.step1 import process_volumes as step1_process_volumes
-from voluseg._steps.step2 import align_volumes as step2_align_volumes
-from voluseg._steps.step3 import mask_volumes as step3_mask_volumes
-from voluseg._steps.step4 import detect_cells as step4_detect_cells
-from voluseg._steps.step5 import clean_cells as step5_clean_cells
+from voluseg.steps.step0 import process_parameters as step0_process_parameters
+from voluseg.steps.step1 import process_volumes as step1_process_volumes
+from voluseg.steps.step2 import align_volumes as step2_align_volumes
+from voluseg.steps.step3 import mask_volumes as step3_mask_volumes
+from voluseg.steps.step4 import detect_cells as step4_detect_cells
+from voluseg.steps.step5 import clean_cells as step5_clean_cells
 
-from voluseg._tools.parameter_dictionary import parameter_dictionary
-from voluseg._tools.parameters import load_parameters, save_parameters
-from voluseg._tools.load_metadata import load_metadata
+from voluseg.tools.parameter_dictionary import parameter_dictionary
+from voluseg.tools.parameters import load_parameters, save_parameters
+from voluseg.tools.load_metadata import load_metadata
 
-from voluseg._update import voluseg_update as update
+from src.voluseg.update import voluseg_update as update
