+++ src/voluseg/tools/sparseness_projection.py
